+++ src/core/enums/ConfigurationKey.ts
@@ -1,3 +1,4 @@
 export enum ConfigurationKey {
   DetectIcons = 'parableWorkspaces.detectIcons',
+  OpenNewWindow = 'parableWorkspaces.openNewWindow',
 }